ayyy, lmao. It works.
Changed String to char[]; it might be a slightly larger buffer size, but it's worth saving CPU instructions ( or something :^) )
Changed around the transferFrom(Peer) method on a lot of Transmittable objects to use a new peer's IN_BUFFER rather than reading from the channel again; we don't need to cause a ReadPendingException.
Fixed some slight "bug" in HashUtils that caused it to run very slowly. Remember to read stuff in chunks and not 1 byte at a time, kids.